access_token_guard: Allow signing in users through guard
Like it is possible with the session guard.

This makes it a lot simpler to use a non-lucid user provider for the access token guard.
Otherwise controllers would have to use "authenticateAsClient" - which is meant for test purposes,
or the token creation logic would have to be extracted to a separate service/provider which would then have to be made available to the controller.